Refacto: vue-pinia in a specific module
